requirements.txt
azure_storage==0.37.0
msal==1.30.0
numpy==1.26.4
openai==1.42.0
Requests==2.32.3
tenacity==9.0.0
tqdm==4.66.5
rich==13.7.1
pyautogen==0.2.35